refactor: ファイルドロップ受け口を Layout に引き上げサイドバーも対象に
- Layout 側で useFileDrop を呼び、AppBar 配下の領域全体（Sidebar +
  main）でドロップを受け付け
- ドロップ時のコンフリクト判定ダイアログ・ドラッグ中オーバーレイ・
  エラー Snackbar を Layout に集約
- TiptapEditor からドロップ関連コードを完全撤去し、エディタ機能のみに
  整理
- useFileDrop API を変更: error state を取りやめ onError コールバック
  方式にして effect 内 setState 警告を解消
